+++ src/air.h
@@ -1,0 +1,11 @@
+#ifndef __AIR_H
+#define __AIR_H
+
+#include <Arduino.h>
+
+#include <Adafruit_VL53L0X.h>
+
+#include "defs.h"
+
+uint8_t air_check(Adafruit_VL53L0X vl53l0x[VL53L0X_COUNT]);
+#endif
