+++ src/report.h
@@ -43,7 +43,7 @@ struct inputdata {
   unsigned char TouchValue[32];  //32个触摸数值
   unsigned char CardStatue;      //读卡器卡片状态，0：无卡，1：aime，2：Felica
   unsigned char CardID[10];      //卡片ID，aime：10字节数据，Felica：8字节数据 + 2字节留空
-};
+} data_tx;
 
 struct rgb {
   unsigned char R;
@@ -54,7 +54,7 @@ struct rgb {
 struct usb_output_data_1 {
   unsigned char Index;       //该字节要求，为0 (即, 接收到的数据如果第1byte为0, 那么就按照usb_output_data_1的结构来解析)
   struct rgb TouchArea[20];  //对应触摸区域前20个灯
-};
+} data_rx_1;
 
 struct usb_output_data_2 {
   unsigned char Index;       //该字节要求，为1 (即, 接收到的数据如果第1byte为1, 那么就按照usb_output_data_2的结构来解析)
@@ -63,6 +63,10 @@ struct usb_output_data_2 {
   struct rgb Rightair;       //右侧板灯光
   struct rgb CardReader;     //读卡器灯光
   unsigned char Empty[18];   //填0以填充UBS规定长度 (因为hid报告描述符里规定下行数据就是61字节所以用0填充到规定长度)
-};
+} data_rx_2;
+
+uint8_t air_check();
+uint16_t get_report_callback(uint8_t report_id, hid_report_type_t report_type, uint8_t* buffer, uint16_t reqlen);
+void set_report_callback(uint8_t report_id, hid_report_type_t report_type, uint8_t const* buffer, uint16_t bufsize);
 
 #endif
